+++ CHANGELOG.md
@@ -6,3 +6,4 @@ v0.2.0
 
 - Added the actual object encontered to the error reported when the
   first element in a tag is not a string.
+- Fixed input structures being modified by AntiHTML.